pyproject.toml
[project]
name = "real-time-rag-application-with-apache-kafka"
version = "0.1.0"
description = "Real-time RAG Pipeline with Kafka + Chroma + Streamlit"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "kafka-python>=2.2.2",
    "requests>=2.32.5",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.25.2",
    "apscheduler>=3.10.4",
    "pydantic>=2.5.3",
    "python-dotenv>=1.0.0",
    "feedparser>=6.0.11",
    "sentence-transformers>=2.3.1",
    "chromadb>=0.4.22",
    "ollama>=0.1.6",
    "streamlit>=1.31.0",
    "plotly>=5.18.0",
    "numpy>=1.26.4",
    "pandas>=2.0.0",
    "pytrends",
    "urllib3<2.0.0",
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"
